python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "go2-sdk"
version = "1.0.0"
description = "Official Python SDK for the Go2 gRPC API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Go2", email = "support@go2.ge"}
]
keywords = ["go2", "smart-links", "grpc", "api", "sdk", "integrations"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "grpcio>=1.59.0",
    "protobuf>=4.25.0",
]

[project.urls]
Homepage = "https://go2.ge"
Documentation = "https://developers.go2.ge"
Repository = "https://github.com/go2-link/sdk"

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.59.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "mypy-protobuf>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["go2_sdk*"]

[tool.mypy]
python_version = "3.8"
strict = true
